# openapi.yaml
openapi: 3.0.3
info:
  title: ProjectBreak - SSR + API Productos
  version: 1.0.0
  description: >
    - SSR (HTML): tienda pública y dashboard admin.
    - API (JSON): endpoints REST de productos bajo /api.
    Auth admin por sesión (cookie connect.sid de express-session) SOLO en rutas SSR protegidas.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: >
        Cookie de sesión de express-session. Se genera tras POST /login.
  schemas:
...
    ProductCreateForm:
      type: object
      required: [Nombre, Descripcion, Categoria, Talla, Precio, image]
      properties:
        Nombre:
          type: string
        Descripcion:
          type: string
        Categoria:
          type: string
          enum: [Ropa, Calzado, Accesorios, Deportiva, Urbana]
        Talla:
          type: string
          enum: [XS, S, M, L, XL, XXL]
        Precio:
          type: number
        image:
          type: string
          format: binary
    ProductUpdateForm:
      type: object
      properties:
        Nombre:
          type: string
        Descripcion:
          type: string
        Categoria:
          type: string
          enum: [Ropa, Calzado, Accesorios, Deportiva, Urbana]
        Talla:
          type: string
          enum: [XS, S, M, L, XL, XXL]
        Precio:
          type: number
        image:
          type: string
          format: binary
...
  responses:
    HtmlOk:
      description: Respuesta SSR (HTML)
      content:
        text/html:
          schema:
            type: string
          example: "<html><body>...</body></html>"
    RedirectToLogin:
      description: Sin sesión → redirige a /login
      headers:
        Location:
          schema:
            type: string
          example: /login
...
paths:
...
  /dashboard:
    get:
      tags: [SSR - Admin]
      summary: SSR - Dashboard admin con listado de productos
      security:
        - cookieAuth: []
      responses:
        "200":
          $ref: "#/components/responses/HtmlOk"
        "302":
          $ref: "#/components/responses/RedirectToLogin"

    post:
      tags: [SSR - Admin]
      summary: SSR - Crea un producto (form + imagen)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ProductCreateForm"
      responses:
        "302":
          description: Crea OK → redirige al detalle del producto en dashboard
          headers:
            Location:
              schema:
                type: string
              example: /dashboard/6985e3960f9dbb3c0e7e2ed7
        "400":
          description: Imagen obligatoria (SSR)
          content:
            text/html:
              schema:
                type: string
        "500":
          description: Error en servidor (SSR)
          content:
            text/html:
              schema:
                type: string
...
  /dashboard/{productId}:
    get:
      tags: [SSR - Admin]
      summary: SSR - Detalle de producto en dashboard
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/HtmlOk"
        "404":
          description: Producto no encontrado (SSR)
          content:
            text/html:
              schema:
                type: string

    put:
      tags: [SSR - Admin]
      summary: SSR - Actualiza un producto (form + imagen opcional)
      description: >
        Se usa con method-override (POST + ?_method=PUT) desde formularios.
        Si se envía nueva imagen, borra la anterior en Cloudinary usando ImagenId.
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ProductUpdateForm"
      responses:
        "302":
          description: Update OK → redirige al detalle del producto
          headers:
            Location:
              schema:
                type: string
              example: /dashboard/6985e3960f9dbb3c0e7e2ed7
        "404":
          description: Producto no encontrado (SSR)
          content:
            text/html:
              schema:
                type: string
        "500":
          description: Error en servidor (SSR)
          content:
            text/html:
              schema:
                type: string
  /dashboard/{productId}/edit:
    get:
      tags: [SSR - Admin]
      summary: SSR - Formulario de edición de producto
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/HtmlOk"
        "404":
          description: Producto no encontrado (SSR)
          content:
            text/html:
              schema:
                type: string
...